Bundle shellcheck in this package
This will ensure npm creates a bin symlink.
It also ensures there is at least _a_ shellcheck available (in case the
download or untar fails).

The version that is downloaded when publishing will be included, but
overwritten (if successful) on install by the user.

This also makes it easier to configure the bin; just setting the untar
directory as the bin directory.

(To keep from versioning shellcheck, it's gitignored)