Creates a placeholder executable for shellcheck
Previously, the install.sh script was run during packing and on install.
This would cause the tarball on npm's registry to include a shellcheck
executable that would (hopefully) be overwritten during installation.

This provided a nice fallback in case the install.sh script failed at
install time. However, it meant that this package's tarball was much
larger than necessary, as it was shipping the linux build of shellcheck.

Also, since the eventual target platform is unknown at pack time, it
meant that the included binary was always for linux.

This is changed now to include a placeholder script instead of the full
shellcheck binary.

The placeholder script serves two purposes:

1) it is an existant (and executable) file that ensures npm will create
   a bin symlink even before the shellcheck binary is downloaded.
2) it prints a message such that if the install.sh script fails, and
   the real binary isn't successfully downloaded, the user will have
   some knowledge that something went wrong.