Only downloads shellcheck during install not pack
Since we can't ship the right binary based on the platform, it doesn't
really make sense to ship one at all. (It was only the right binary for
linux users.)

Also, it made this package's tarball larger than necessary.

Now we're reverted back to only fetching the shellcheck binary during
installation (preinstall, to be precise)